+++ gr/gr_demod_bpsk_sdr.h
@@ -54,7 +54,7 @@ public slots:
     void stop();
     std::vector<unsigned char> *getData();
     void tune(long center_freq);
-    void set_rx_sensitivity(int value);
+    void set_rx_sensitivity(float value);
 
 private:
     gr::top_block_sptr _top_block;
