+++ gr/gr_demod_nbfm_sdr.cpp
@@ -95,7 +95,7 @@ void gr_demod_nbfm_sdr::tune(long center_freq)
     _osmosdr_source->set_center_freq(_device_frequency-25000);
 }
 
-void gr_demod_nbfm_sdr::set_rx_sensitivity(int value)
+void gr_demod_nbfm_sdr::set_rx_sensitivity(float value)
 {
     osmosdr::gain_range_t range = _osmosdr_source->get_gain_range();
     if (!range.empty())
